Move spatial and temporal disaggregation to own module
